requirements.txt
yfinance
fastapi[all]
google-auth
google-genai